refactor: WebUIChessApplication 리팩토링
