+++ src/main/java/chess/utils/JsonTransformer.java
@@ -1,4 +1,4 @@
-package chess.view;
+package chess.utils;
 
 import com.google.gson.Gson;
 import spark.ResponseTransformer;
