+++ src/main/resources/templates/index.html
@@ -3,14 +3,85 @@
 <head>
     <meta charset="UTF-8">
     <title>체스 게임</title>
+    <style>
+        .container {
+
+        }
+        .title {
+            text-align: center;
+        }
+        .menu_container {
+            padding: 20px;
+            text-align: center;
+        }
+        .game_manual_link {
+            margin: 10px;
+            padding: 20px;
+            font-size: 20px;
+        }
+        body {
+            background-image: url(http://p1.pichost.me/i/11/1344899.jpg);
+            background-size: cover;
+            background-repeat: no-repeat;
+            font-family: Arial, sans-serif;
+            font-weight: bold;
+            font-size: 14px;
+        }
+
+        .wrap {
+            position: absolute;
+            top: 50%;
+            left: 50%;
+            margin-top: -86px;
+            margin-left: -89px;
+            text-align: center;
+        }
+
+        a {
+            -webkit-transition: all 200ms cubic-bezier(0.390, 0.500, 0.150, 1.360);
+            -moz-transition: all 200ms cubic-bezier(0.390, 0.500, 0.150, 1.360);
+            -ms-transition: all 200ms cubic-bezier(0.390, 0.500, 0.150, 1.360);
+            -o-transition: all 200ms cubic-bezier(0.390, 0.500, 0.150, 1.360);
+            transition: all 200ms cubic-bezier(0.390, 0.500, 0.150, 1.360);
+            display: block;
+            margin: 20px auto;
+            max-width: 180px;
+            text-decoration: none;
+            border-radius: 4px;
+            padding: 20px 30px;
+        }
+
+        a.button {
+            color: rgba(30, 22, 54, 0.6);
+            box-shadow: rgba(30, 22, 54, 0.4) 0 0px 0px 2px inset;
+        }
+
+        a.button:hover {
+            color: rgba(255, 255, 255, 0.85);
+            box-shadow: rgba(30, 22, 54, 0.7) 0 0px 0px 40px inset;
+        }
+
+        a.button2 {
+            color: rgba(30, 22, 54, 0.6);
+            box-shadow: rgba(30, 22, 54, 0.4) 0 0px 0px 2px inset;
+        }
+
+        a.button2:hover {
+            color: rgba(255, 255, 255, 0.85);
+            box-shadow: rgba(30, 22, 54, 0.7) 0 80px 0px 2px inset;
+        }
+
+    </style>
 </head>
 <body>
-    <h1>체스 게임에 오신 것을 환영합니다.</h1>
-    <div>
-        <a href="/game_play">새로운 게임 시작</a>
-    </div>
-    <div>
-        <a href="/game_continue">지난 게임 계속하기</a>
+    <div class="container">
+        <div class="menu_container">
+            <h1 class="title">체스 게임에 오신 것을 환영합니다.</h1>
+            <div class="wrap">
+                <a href="/game_play" class="button">새로운 게임 시작</a>
+                <a href="/game_continue" class="button2">지난 게임 계속하기</a>
+            </div>
+        </div>
     </div>
 </body>
 </html>
